+++ src/main.rs
@@ -26,6 +26,7 @@ fn inner() -> tricorder::domain::Result<ExitCode> {
         Command::Pitstop(args) => tricorder::commands::pitstop(&args),
         Command::Postedit(args) => tricorder::commands::post_edit(&args),
         Command::Precommit(args) => tricorder::commands::precommit(&args),
+        Command::Test(args) => tricorder::commands::test(&args),
         Command::UpdateTools => tricorder::commands::update_tools(),
     }
 }